Change project file layout.
The command is now in the root of the repository, with packages being
subdirectories (even if they have the same name).  This simplifies
naming by allowing the command and the package to have the same name.

This commit also presumes the github repository will be renamed from
"go-daycount" to "daycount".